refactor(errors): update error file by error folder with resource error file
