+++ src/domain/file/domain.ts
@@ -18,7 +18,7 @@ import type { ForgetType } from '@/util/type'
 import { isDevMode } from '@/util/env.util'
 import type { FileDescriptor, Query } from './query'
 import type { Eq } from 'fp-ts/lib/Eq'
-import { ResourceAlreadyExistsError, ResourceNotFoundError } from '@/shared/error'
+import { ResourceAlreadyExistsError, ResourceNotFoundError } from '@/shared/error/resource'
 
 export type State = {
   data: File[]
