+++ src/shared/error/resource.ts
@@ -24,7 +24,7 @@ export type ResourceNotFoundError = ReturnType<typeof ResourceNotFoundError>
 
 export type ResourceError = ResourceAlreadyExistsError | ResourceNotFoundError
 
-export const ShowFileError: Show<ResourceError> = {
+export const ShowResourceError: Show<ResourceError> = {
   show: (error: ResourceError): string => {
     switch (error._tag) {
       case 'resource-already-exists': {
